third_party/WebKit/LayoutTests/typedcssom/styleValue-stringifier.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>

var styleValues = {
  // TODO(meade): Add each CSSStyleValue type here.
  'initial': new CSSKeywordValue('initial'),
  '10px': new CSSUnitValue(10, 'px'),
  'rotate(10deg)': new CSSTransformValue(
      [new CSSRotation(new CSSUnitValue(10, 'deg'))]),
};

for (var cssText in styleValues) {
  test(function() {
    assert_equals(styleValues[cssText].toString(), cssText);
    assert_equals(styleValues[cssText].toString(), cssText);
  }, 'Style value with type ' + styleValues[cssText].constructor.name +
      'with value "' + cssText + '" serializes correctly');
}

</script>
