third_party/WebKit/LayoutTests/typedcssom/styleValue-stringifier-expected.txt

This is a testharness.js-based test.
PASS Style value with type CSSKeywordValuewith value "initial" serializes correctly 
PASS Style value with type CSSUnitValuewith value "10px" serializes correctly 
FAIL Style value with type CSSTransformValuewith value "rotate(10deg)" serializes correctly assert_equals: expected "rotate(10deg)" but got ""
Harness: the test ran to completion.
